requirements.txt
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
iniconfig==2.0.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
PyMySQL==1.1.0
pytest==8.1.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.4.0
urllib3==2.2.1
WooCommerce==3.0.0
